Proyecto Final/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash-ag-grid>=31.2.0",
    "dash-svg>=0.0.12",
    "dash>=2.18.2",
    "langchain-openai>=0.2.8",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
